fix migration integer problem

+++ database/migrations/2016_04_14_000001_create_users_table.php
@@ -13,13 +13,13 @@ class CreateUsersTable extends Migration
     public function up()
     {
         Schema::create('users', function (Blueprint $table) {
-            $table->increments('id');
+            $table->bigIncrements('id');
             $table->string('first_name');
             $table->string('last_name');
             $table->string('email')->unique();
             $table->string('password');
             $table->dateTime('last_login')->nullable();
-            $table->integer('faculty_id')->unsigned();
+            $table->bigInteger('faculty_id')->unsigned();
             $table->enum('gender', ['Kadın', 'Erkek'])->nullable();
             $table->date('birthday');
             $table->string('mobile', 10);
