+++ database/migrations/2018_11_16_161046_add_stop_to_faculty.php
@@ -15,7 +15,7 @@ class AddStopToFaculty extends Migration
     {
         Schema::table('faculties', function (Blueprint $table) {
             $table->dateTime('stopped_at')->nullable();
-            $table->integer('stopped_by')->unsigned()->nullable();
+            $table->bigInteger('stopped_by')->unsigned()->nullable();
             $table->foreign('stopped_by')->references('id')->on('users');
         });
     }
